requirements.txt
opencv-python
numpy
fastapi
uvicorn
python-multipart
Pillow
rembg